refactored and added testfiles for at and tt
