+++ packages/frontend/tests/performance/k6/commonFunctions/getCookie.js
@@ -1,5 +1,37 @@
 import http from 'k6/http';
 import { afUrl } from './config.js';
+import { getPersonalToken } from './token.js';
+
+const environment = __ENV.ENVIRONMENT || 'yt';
+
+const tokenGeneratorEnv = (() => {
+  switch (environment) {
+      case 'at':
+          return 'at23';
+      case 'tt':
+          return 'tt02';
+      case 'yt':
+          return 'yt01';
+      default:
+          return 'yt01';
+  }
+}) ();
+
+/**
+ * Function to get a personal token for a given PID.
+ * @param {string} pid - The personal identification number (PID) of the user.
+ * @return {string} - The generated personal token.
+ **/
+function getToken(pid) {
+    const tokenParams = {
+      scopes: "digdir:dialogporten.noconsent openid altinn:portal/enduser",
+      pid: pid,
+      env: tokenGeneratorEnv
+    }
+    const token = getPersonalToken(tokenParams);
+    return token
+  }
+
 
 function getSessionId(token) {
     const url = new URL(`${afUrl}/api/init-session`);
@@ -22,7 +54,8 @@ function getSessionId(token) {
     return sessionId // Replace with actual logic to get session ID
 }
 
-export function getCookie(token) {
+export function getCookie(pid) {
+    var token = getToken(pid);
     const cookie = {
         name: "arbeidsflate",
         value: getSessionId(token),
