+++ packages/frontend/tests/performance/k6/commonFunctions/token.js
