--- packages/frontend/tests/performance/k6/conf-browser.yaml
@@ -1,13 +1,0 @@
-namespace: dialogporten
-test_definitions:
-  - test_file: packages/frontend/tests/performance/k6/browser-no-cookie.js
-    contexts:
-      - environment: yt01
-        node_type: "default"
-        test_type:
-          type: browser
-          enabled: true
-        test_run:
-          env:
-            - name: ENVIRONMENT
-              value: "yt01"
