--- packages/frontend/tests/performance/k6/conf.yaml
@@ -1,17 +1,0 @@
-namespace: dialogporten
-test_definitions:
-  - test_file: packages/frontend/tests/performance/k6/browser.js
-    contexts:
-      - environment: yt01
-        node_type: "default"
-        test_type:
-          type: browser
-          enabled: true
-        test_run:
-          env:
-            - name: ENVIRONMENT
-              value: "yt01"
-          secrets:
-            - redis-connection-information
-            - token-generator-creds
-            - signature-credentials
