+++ packages/frontend/tests/performance/k6/testData/usersWithDialogs-at.csv
@@ -26,3 +26,13 @@ pid
 57897700966
 07855499058
 12918698416
+11916497514
+10830097746
+18865296564
+21817795517
+03815196960
+21908397883
+21918996823
+05917698181
+20897196558
+19874895949
