+++ packages/frontend/tests/performance/k6/tests/browser.js
@@ -1,0 +1,63 @@
+
+
+/**
+ * Async function to select a side menu element on a page.
+ * @param {object} page - The page object to interact with.
+ * @param {string} locator - Locator for the menu element.
+ * @param {object} trend - Trend metric to track the action duration.
+ */
+export async function selectSideMenuElement(page, locator, trend) {
+  var menuElement = await page.waitForSelector(locator, { timeout: 2000 }).catch(() => false);
+  var startTime = new Date();
+  await Promise.all([
+    menuElement.click(),
+  ]);
+
+  // Wait for the page to load after clicking the menu element
+  await waitForPageLoaded(page);
+  // Track the time taken for the action
+  var endTime = new Date();
+  trend.add(endTime - startTime);
+}
+
+/**
+ * Async function to select the next page in the dialog list.
+ * @param {object} page - The page object to interact with.
+ * @param {object} trend - Trend metric to track the action duration.
+ * @return {Promise<void>} - A promise that resolves when the next page is selected.
+ */
+export async function selectNextPage(page, trend) {
+  var next_page = await page.waitForSelector('button[class="ds-button"]', { state: 'attached', timeout: 500 }).catch(() => false);
+  var iterations = 0;
+  while (next_page && iterations < 10) {
+    var startTime = new Date();
+    await Promise.all([
+      next_page.click(),
+    ]);
+   
+    // Wait for the page to load after clicking the next page button
+    await waitForPageLoaded(page, 2);
+    next_page = await page.waitForSelector('button[class="ds-button"]', { state: 'attached', timeout: 500 }).catch(() => false);
+    var endTime = new Date();
+    trend.add(endTime - startTime);
+    iterations++;
+  }
+}
+
+/**
+ * Async function to wait for the page to load.
+ * @param {object} page - The page object to interact with.
+ * @param {number} empties - Number of empty checks to perform (default is 1).
+ * @return {Promise<void>} - A promise that resolves when the page is loaded.
+ */
+async function waitForPageLoaded(page, empties = 1) {
+  var busyItems = await page.$$('li [aria-busy="true"]');
+  var noEmptys = 0;
+  while ( busyItems.length > 0 || noEmptys < empties) {
+    await page.waitForTimeout(10); // Wait for 10 ms before checking again
+    busyItems = await page.$$('li [aria-busy="true"]');
+    if (busyItems.length == 0) {
+      noEmptys++;
+    }
+  }
+}
